Step 8 (partial): extract navigation.js, kill toggleDspKey duplicate
Closes cycle 5 (controls.js <-> renderer.js, per CLAUDE.md's
canonical cycle numbering). Removing the dead `keypressMasks`
import from renderer.js:5 eliminates the only renderer->controls
edge; the residual controls->renderer edge (updateScreen, called
after a keypress) is one-way and no longer a cycle. Full removal
of that residual coupling would require routing the post-keypress
refresh through events.js, deferred.

The toggleDspKey duplicate at controls.js:65 and renderer.js:37 —
flagged as a sub-finding in docs/refactor/02-top-couplings.md §4 —
is the concrete symptom this commit removes; navigation.js becomes
its canonical home and all three external call sites (main.js:151,
controls.js:128, tests/renderer.test.js) import from there.

Also drops the dead `keypressMasks` import from renderer.js:5
(flagged in 02-top-couplings.md §4 as "likely dead import" —
confirmed by grep, no usage sites in renderer.js).

Renderer folder-split deliberately deferred per roadmap
docs/refactor/04-roadmap.md L101 exit ramp: the existing 3-branch
renderer snapshot gate is insufficient for a renderer split (does
not pin graphic-EQ position-'a' grouping, embedded childSubs,
keyStack depth >2, SET hex/dec index >=10, or INF). Widening that
gate is a session of work in itself and the folder split delivers
no cycle closure. Step 8 as committed = navigation extraction only.

tests/startup.test.js:133 contains an intentional local re-definition
of toggleDspKey (see comment at :130); left untouched per its
documented purpose of keeping the startup simulation standalone.
