+++ src/main.js
@@ -3,7 +3,8 @@ import { WebMidi } from 'webmidi';
 import { loadConfig, saveConfig, clearConfig } from './config.js';
 import { setupKeypressControls, testKeypress } from './controls.js';
 import { setMidiPorts, addSysexListener, sendSysEx, sendValueDump, sendValuePut, sendObjectInfoDump } from './midi.js';
-import { updateScreen, toggleDspKey } from './renderer.js';
+import { updateScreen } from './renderer.js';
+import { toggleDspKey } from './navigation.js';
 import { appState } from './state.js';
 import { setState } from './store.js';
 import { denibble, renderBitmap } from './bitmap.js';
