+++ src/renderer.js
@@ -2,7 +2,6 @@
 import { appState } from './state.js';
 import { setState } from './store.js';
 import { sendObjectInfoDump, sendValueDump, sendValuePut, sendSysEx } from './midi.js';
-import { keypressMasks } from './controls.js';
 import { showLoading } from './main.js';
 import { log } from './logger.js';
 
@@ -23,19 +22,6 @@ export function updateScreen(logParam = null) {
   setState(patch, 'renderer:update-screen-clear');
   sendObjectInfoDump(appState.currentKey, logParam);
   sendValueDump(appState.currentKey, logParam);
-}
-
-/**
- * Toggles a DSP key between '4' (A) and '8' (B) prefixes.
- * 
- * @param {string} key - The DSP key to toggle (e.g., '401000b').
- * @returns {string} The toggled key (e.g., '801000b').
- * 
- * @example
- * toggleDspKey('401000b'); // '801000b'
- */
-export function toggleDspKey(key) {
-  return key.startsWith('4') ? '8' + key.slice(1) : '4' + key.slice(1);
 }
 
 /**
